requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
deepface==0.0.72
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-import-export==2.7.1
django-widget-tweaks==1.4.8
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.4.2
Flask==2.0.2
flatbuffers==2.0
fontawesome-free==5.15.4
gast==0.4.0
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mtcnn==0.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.4
oauthlib==3.1.1
odfpy==1.4.1
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
prometheus-client==0.13.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pywin32==303
pywinpty==2.0.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.9
rsa==4.8
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
stack-data==0.1.4
tablib==3.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.13.0
testpath==0.5.0
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
